backend/hostinger_deployment/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=1.9.0,<2.0.0
python-dotenv==1.0.0
openai==1.3.7
httpx>=0.23.0,<0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
supabase==1.0.3
postgrest-py==0.10.3
gunicorn==21.2.0
uvicorn[standard]==0.25.0
